backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "luiz guedes",email = "pitohuikun@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12" 
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.9)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    # --- NOVAS DEPENDÊNCIAS DO SISTEMA ASSÍNCRONO ---
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "django-celery-results (>=2.5.0,<3.0.0)",
    # Scrapers (app `sources`)
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "requests (>=2.32.0,<3.0.0)"
]

[tool.poetry]
packages = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"